Create squeeze screener_v2
This document contains the body of the screener.
It firstly sets up the parameters of the Bollinger Band and Keltner Channel. Then, I define the in squeeze condition. Choosing stocks from the list, for example, it is the S&P 500 components in this case, and screen if there is any stocks matching the condition. If so, it would says" {symbol} has started squeezing and show you the figure of its price movement